+++ biceps/src/main/java/org/somda/sdc/biceps/provider/access/LocalMdibAccessImpl.java
@@ -136,6 +136,11 @@ public class LocalMdibAccessImpl implements LocalMdibAccess {
     }
 
     @Override
+    public void unregisterAllObservers() {
+        eventDistributor.unregisterAllObservers();
+    }
+
+    @Override
     public MdibVersion getMdibVersion() {
         try (ReadTransaction transaction = startTransaction()) {
             return transaction.getMdibVersion();