+++ common/src/main/java/org/somda/sdc/common/guice/DefaultCommonModule.java
@@ -1,6 +1,8 @@
 package org.somda.sdc.common.guice;
 
 import com.google.inject.AbstractModule;
+import org.somda.sdc.common.event.EventBus;
+import org.somda.sdc.common.event.EventBusImpl;
 import org.somda.sdc.common.util.JaxbUtil;
 import org.somda.sdc.common.util.JaxbUtilImpl;
 
@@ -11,5 +13,6 @@ public class DefaultCommonModule extends AbstractModule {
     @Override
     protected void configure() {
         bind(JaxbUtil.class).to(JaxbUtilImpl.class).asEagerSingleton();
+        bind(EventBus.class).to(EventBusImpl.class);
     }
 }